pyproject.toml
[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[project]
name = "magma"
version = "0.0.1"
description = "A Foundation Model for Multimodal AI Agents."
readme = "README.md"
requires-python = ">=3.10"
classifiers = [
    "Programming Language :: Python :: 3",
    "License :: OSI Approved :: Apache Software License",
]
dependencies = [
    "torch",
    "torchvision", 
    "pytorch-lightning",
    "transformers",
    "tokenizers",
    "sentencepiece",
    "shortuuid",
    "accelerate",
    "peft",
    "bitsandbytes",
    "pydantic",
    "numpy",
    "scikit-learn",
    "gradio",
    "gradio_client",
    "spaces",
    "requests",
    "httpx",
    "uvicorn",
    "fastapi",
    "einops",
    "einops-exts",
    "timm",
    "tensorflow",
    "tensorflow_datasets",
    "tensorflow_graphics",    
    "draccus",    
    "pyav",
    "numba", 
    "dlimp @ git+https://github.com/moojink/dlimp_openvla",
    "loguru", 
    "sacrebleu", 
    "evaluate", 
    "sqlitedict",
    "open_clip_torch",
    "flash-attn",
    "datasets==2.14.6"
]

[project.optional-dependencies]
train = [
    "deepspeed", 
    "ninja", 
    "wandb"
]
eval = [
    "azure-ai-ml",
    "datasets",
    "fire",
    "openai",
    "opencv-python",
    "openpyxl",
    "pillow",
    "python-Levenshtein",
    "rich",
    "streamlit",
    "typer[all]",
    "word2number",
]
agent = [
    "pygame",
    "easyocr",
    "paddleocr",
    "common==0.1.2",
    "dual==0.0.10",
    "tight==0.1.0",
    "prox==0.0.17",
    "paddle==1.0.2",
    "paddlepaddle==2.6.2",
    "supervision==0.18.0",
    "ultralytics==8.3.78", 
]

[tool.setuptools.packages.find]
exclude = [
    "assets",
    "benchmark",
    "docs",
    "dist",
    "playground",
    "scripts",
    "tests",
    "azureblobs",
    "azure"
]
[tool.wheel]
exclude = [
    "assets",
    "benchmark",
    "docs",
    "dist",
    "playground",
    "scripts",
    "tests",
    "azureblobs",
    "azure"    
]
[tool.black]
line-length = 120
skip-string-normalization = true
[tool.pyright]
exclude = [
    "**/pycache",
    "playground",
    "_results",
    "_data",
    "models",
    "checkpoints",
    "wandb",
    "docs",
]